# Conflicts:
#	Engine/Config/BaseEditorSettings.ini
#	Engine/Config/BaseEngine.ini
#	Engine/Source/Runtime/D3D12RHI/Private/D3D12Viewport.cpp
#	Engine/Source/Runtime/Slate/Private/Widgets/Text/SlateEditableTextLayout.cpp

+++ Engine/Source/Runtime/OpenGLDrv/Private/Mac/OpenGLMac.cpp
@@ -1,4 +1,4 @@
-// Copyright 1998-2016 Epic Games, Inc. All Rights Reserved.
+// Copyright 1998-2017 Epic Games, Inc. All Rights Reserved.
 
 #include "OpenGLDrvPrivate.h"
 
@@ -756,7 +756,7 @@ bool PlatformGetAvailableResolutions(FScreenResolutionArray& Resolutions, bool b
 	if (AllModes)
 	{
 		const int32 NumModes = CFArrayGetCount(AllModes);
-		const int32 Scale = FMacApplication::GetPrimaryScreenBackingScaleFactor();
+		const int32 Scale = (int32)FMacApplication::GetPrimaryScreenBackingScaleFactor();
 
 		for (int32 Index = 0; Index < NumModes; Index++)
 		{
